+++ packages/terra-button/src/terra-dev-site/test/button/ButtonTypes.test.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 
 import Button from '../../../Button';
+import styles from './ButtonTestCommon.module.scss';
 
-const spacingStyle = { padding: '20px' };
+const cx = classNames.bind(styles);
 
 const ButtonWithTypeButton = () => <Button id="buttonWithTypeButton" text="Button Type is Default" type="button" />;
 const ButtonWithTypeReset = () => <Button id="buttonWithTypeReset" text="Button Type is Reset" type="reset" />;
@@ -11,9 +13,9 @@ const ButtonWithTypeSubmit = () => <Button id="buttonWithTypeSubmit" text="Butto
 export default () => (
   <div>
     <ButtonWithTypeButton />
-    <div style={spacingStyle} />
+    <div className={cx('spacing')} />
     <ButtonWithTypeReset />
-    <div style={spacingStyle} />
+    <div className={cx('spacing')} />
     <ButtonWithTypeSubmit />
   </div>
 );