+++ packages/terra-form-field/src/terra-dev-site/test/form-field/FieldHiddenLabel.test.jsx
@@ -1,16 +1,20 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Field from '../../../Field';
+import styles from './FormFieldTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const FieldHiddenLabel = () => (
   <div>
     <h3>Field - Hidden Label</h3>
     <Field
       id="label-hidden"
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       isLabelHidden
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
   </div>
 );