+++ packages/terra-form-field/src/terra-dev-site/test/form-field/FieldInteractiveInvalid.test.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Field from '../../../Field';
+import styles from './FormFieldTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 class FieldExamples extends React.Component {
   constructor(props) {
@@ -16,48 +20,48 @@ class FieldExamples extends React.Component {
     const field = (
       <Field
         id="default"
-        style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+        className={cx('form-field')}
         label="Field Label"
         help="Help Message"
         error="Error Message"
         isInvalid={this.state.isInvalid}
       >
-        <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+        <div className={cx('field-content')}>Control Placeholder</div>
       </Field>
     );
 
     const optional = (
       <Field
         id="optional"
-        style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+        className={cx('form-field')}
         label="Field Label"
         help="Help Message"
         error="Error Message"
         isInvalid={this.state.isInvalid}
         showOptional
       >
-        <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+        <div className={cx('field-content')}>Control Placeholder</div>
       </Field>
     );
 
     const required = (
       <Field
         id="required"
-        style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+        className={cx('form-field')}
         label="Field Label"
         help="Help Message"
         error="Error Message"
         isInvalid={this.state.isInvalid}
         required
       >
-        <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+        <div className={cx('field-content')}>Control Placeholder</div>
       </Field>
     );
 
     const requiredHidden = (
       <Field
         id="required-hidden"
-        style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+        className={cx('form-field')}
         label="Field Label"
         help="Help Message"
         error="Error Message"
@@ -65,21 +69,21 @@ class FieldExamples extends React.Component {
         required
         hideRequired
       >
-        <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+        <div className={cx('field-content')}>Control Placeholder</div>
       </Field>
     );
 
     const hiddenLabel = (
       <Field
         id="label-hidden"
-        style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+        className={cx('form-field')}
         label="Field Label"
         help="Help Message"
         error="Error Message"
         isInvalid={this.state.isInvalid}
         isLabelHidden
       >
-        <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+        <div className={cx('field-content')}>Control Placeholder</div>
       </Field>
     );
 
@@ -87,19 +91,19 @@ class FieldExamples extends React.Component {
       <div>
         <Field
           id="inline-1"
-          style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+          className={cx('form-field')}
           label="Field Label - Spacing POC"
           error="Error Message"
           isInline
           isInvalid={this.state.isInvalid}
           required
         >
-          <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+          <div className={cx('field-content')}>Control Placeholder</div>
         </Field>
 
         <Field
           id="inline-2"
-          style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+          className={cx('form-field')}
           label="Field Label - Spacing POC"
           error="Error Message"
           isInline
@@ -107,19 +111,19 @@ class FieldExamples extends React.Component {
           required
           hideRequired
         >
-          <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+          <div className={cx('field-content')}>Control Placeholder</div>
         </Field>
 
         <Field
           id="inline-3"
-          style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+          className={cx('form-field')}
           label="Field Label - Spacing POC"
           error="Error Message"
           isInline
           isInvalid={this.state.isInvalid}
           showOptional
         >
-          <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+          <div className={cx('field-content')}>Control Placeholder</div>
         </Field>
       </div>
     );