+++ packages/terra-form-field/src/terra-dev-site/test/form-field/FieldLabelErrorHelpText.test.jsx
@@ -1,18 +1,22 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import Field from '../../../Field';
+import styles from './FormFieldTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
 
 const FieldLabelErrorHelpText = () => (
   <div>
     <h3>Field - Label + In Error + Help Text</h3>
     <Field
       id="optional-invalid"
-      style={{ border: 'dashed 1px lightGrey', padding: '10px' }}
+      className={cx('form-field')}
       label="Field Label"
       help="Help Message"
       error="Error Message"
       isInvalid
     >
-      <div style={{ border: 'dashed 1px lightGrey' }}>Control Placeholder</div>
+      <div className={cx('field-content')}>Control Placeholder</div>
     </Field>
   </div>
 );